InstPackages* scripts to build packages for three kinds of releases.
At the moment they are identical for all three branches.
However (and it was happening in the past) there are situations
when build instructions are changing and different versions of
the package should be used for update/minor/major release. Hence
we need to provide three different versions of InstPackages*.sh.

After the release, one should publish appropriate versions of
these scripts on the GAP website:
https://github.com/gap-system/GapWWW/tree/master/Download