Withdrawing QaoS package from the GAP package distribution
Quoting https://mail.gap-system.org/pipermail/gap/2019-March/001155.html

Unfortunately we have to withdraw the QaoS package from distribution of GAP,
as the servers it crucially relies on for its functionality have been
permanently retired some time ago and are not coming back (see
<https://github.com/gap-packages/qaos/issues/13> for details). Without them,
this package is mostly useless.

Of course the source code repository on GitHub will stay around, just in
case somebody is interested in the code for other reasons.