DistributionUpdate: remove 'releasebranch' handling
The three-step process in the `doit` script begun by creating a
"releasebranch" directory, which was turned into an archive, and then
extracted under a different name, `gapcore`. The only effect this had
was to exclude a few files based on the pattern*.txt files.  However, we
exclude these files when creating the archive from `gapcore`, too, so in
the end, there seems to be no reason for this intermediate step.

Hence, instead of doing that, just create the initial directory directly
in the final destination with its final name `gapcore`.